fix #42659, move `jl_coverage_visit_line` to runtime library
