+++ src/julia.expmap
@@ -37,6 +37,9 @@
     jlbacktracet;
     _IO_stdin_used;
     __ZN4llvm23createLowerSimdLoopPassEv;
+    _Z24jl_coverage_data_pointerN4llvm9StringRefEi;
+    _Z22jl_coverage_alloc_lineN4llvm9StringRefEi;
+    _Z22jl_malloc_data_pointerN4llvm9StringRefEi;
     LLVMExtra*;
 
     /* freebsd */
